Refactored experiments into a class and refactored associated utilities
Moved default hyperparameter calculations for optimizers to experiment class
